refactor: move all validation functions into utils
