+++ src/components/OrganisationEditor/index.tsx
@@ -1,10 +1,10 @@
+import { validatePostalCode } from "@/src/utils/validation";
 import _ from "lodash";
 import { FC, FormEvent, useEffect, useState } from "react";
 import { CreateOrganizationRequest } from "../../api/client/models/CreateOrganizationRequest";
 import { Organization } from "../../api/client/models/Organization";
 import { Button } from "../../components/Button";
 import { Input } from "../../components/InputField";
-import { validatePostalCode } from "../../utils/validation";
 
 interface ErrorMessages {
 	postalCode: string | undefined;
