+++ src/utils/validation.tsx
@@ -2,3 +2,17 @@ export const validatePostalCode = (value: string): string => {
 	const isValid = /^\d{5}$/.test(value);
 	return isValid ? "" : "PLZ muss aus fünf Zahlen bestehen";
 };
+
+export const validateEmail = (email: string): string => {
+	return email.match(/^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/)
+		? ""
+		: "Bitte gib eine gültige E-Mail Adresse ein";
+};
+
+export const validatePassword = (password: string): string => {
+	return password.length >= 12 ? "" : "Bitte gib ein Passwort mit mindestens 8 Zeichen ein";
+};
+
+export const validateRepeatPassword = (password: string, passwordRepeat: string): string => {
+	return password === passwordRepeat ? "" : "Passwörter stimmen nicht überein";
+};
